refactor!: read mesh in meshRegularizer
