+++ src/CMakeLists.txt
@@ -18,6 +18,7 @@ set(DDG_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/solver/pressure.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/solver/energy.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/solver/regularization.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/solver/mesh_process.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/solver/trajfile.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/solver/mutable_trajfile.cpp"
