+++ src/solver/integrator/conjugate_gradient.cpp
@@ -69,7 +69,7 @@ bool ConjugateGradient::integrate() {
     // Process mesh every tProcessMesh period
     if (f.time - lastProcessMesh > tProcessMesh) {
       lastProcessMesh = f.time;
-      f.processMesh();
+      f.mutateMesh();
       f.updateVertexPositions(false);
     }
