+++ src/solver/integrator/stormer_verlet.cpp
@@ -67,7 +67,7 @@ bool StormerVerlet::integrate() {
     }
     // std::cout << "total force:  " << totalForce.norm() << std::endl;
     // process the mesh with regularization or mutation
-    f.processMesh();
+    f.mutateMesh();
     f.updateVertexPositions();
     pastPositions = temp;
     // initialize/update the vertex position of the last
